+++ src/js/core/transform/Draggable.js
@@ -1,7 +1,7 @@
 import { helper } from '../Helper';
 import Transformable from './Transformable';
-import { floatToFixed, getMinMaxOfArray } from './common';
-import { isDef, isUndef } from '../util/util';
+import { floatToFixed, getMinMaxOfArray, DEG } from './common';
+import { isDef, isUndef, warn } from '../util/util';
 import { addClass, matrixToCSS, getScrollOffset, getElementOffset } from '../util/css-util';
 import { MIN_SIZE, CLIENT_EVENTS_CONSTANTS } from '../consts';
 
@@ -30,12 +30,13 @@ export default class Draggable extends Transformable {
     _init(elements) {
         const {
             options: {
-                rotationPoint,
+                transformOrigin,
                 container,
                 controlsContainer,
                 resizable,
                 rotatable,
-                showNormal
+                showNormal,
+                restrict
             }
         } = this;
 
@@ -61,7 +62,7 @@ export default class Draggable extends Transformable {
 
             let radius = null;
 
-            if (rotationPoint) {
+            if (transformOrigin) {
                 radius = renderLine([finalVertices.center, finalVertices.center], 'radius');
                 addClass(radius, 'sjx-hidden');
 
@@ -95,14 +96,14 @@ export default class Draggable extends Transformable {
             }
             : {};
 
-        const nextCenter = Array.isArray(rotationPoint)
-            ? [...rotationPoint, 0, 1]
+        const nextTransformOrigin = Array.isArray(transformOrigin)
+            ? [...transformOrigin, 0, 1]
             : finalVertices.center;
 
         const allHandles = {
             ...resizingHandles,
-            center: rotationPoint && rotatable
-                ? nextCenter
+            center: transformOrigin && rotatable
+                ? nextTransformOrigin
                 : undefined,
             rotator
         };
@@ -137,6 +138,8 @@ export default class Draggable extends Transformable {
             })
         ));
 
+        const restrictContainer = restrict || container;
+
         this.storage = {
             wrapper,
             controls,
@@ -146,8 +149,13 @@ export default class Draggable extends Transformable {
             },
             data,
             center: {
-                isShifted: Array.isArray(rotationPoint)
-            }
+                isShifted: Array.isArray(transformOrigin)
+            },
+            transformOrigin: nextTransformOrigin,
+            transform: {
+                containerMatrix: getCurrentTransformMatrix(restrictContainer, restrictContainer.parentNode)
+            },
+            cached: {}
         };
 
         [...elements, controls].map(target => (
@@ -272,9 +280,9 @@ export default class Draggable extends Transformable {
             $el.css(css);
             $controls.css(css);
         }
-
-        this.storage.cached = {};
     }
+
+    _processActions() { }
 
     _processResize(element, { dx, dy }) {
         const {
@@ -406,8 +414,7 @@ export default class Draggable extends Transformable {
     _processMove(element, { dx, dy }) {
         const {
             storage: {
-                data,
-                center
+                data
             }
         } = this;
 
@@ -451,16 +458,6 @@ export default class Draggable extends Transformable {
                 }
             }
         });
-
-        if (center.isShifted) {
-            // const centerTransformMatrix = dropTranslate(matrixInvert(wrapperMatrix));
-            // const [cx, cy] = multiplyMatrixAndPoint(
-            //     centerTransformMatrix,
-            //     [nextDx, nextDy, 0, 1]
-            // );
-
-            // this._moveCenterHandle(-cx, -cy);
-        }
 
         return moveElementMtrx;
     }
@@ -518,12 +515,11 @@ export default class Draggable extends Transformable {
                 handles: {
                     center: cHandle
                 },
-                wrapper,
-                data
+                data,
+                transformOrigin
             },
             options: {
                 container,
-                restrict,
                 scalable
             }
         } = this;
@@ -531,11 +527,6 @@ export default class Draggable extends Transformable {
         const storage = data.get(element);
 
         const { parent } = storage;
-
-        const {
-            offsetWidth,
-            offsetHeight
-        } = restrict || container;
 
         const [glLeft, glTop] = getAbsoluteOffset(element, container);
 
@@ -549,10 +540,6 @@ export default class Draggable extends Transformable {
         const matrix = getTransform(element);
         const ctm = getCurrentTransformMatrix(element, container);
         const parentMatrix = getCurrentTransformMatrix(parent, container);
-        const wrapperMatrix = getCurrentTransformMatrix(wrapper, container);
-        const containerMatrix = restrict
-            ? getCurrentTransformMatrix(restrict, restrict.parentNode)
-            : getCurrentTransformMatrix(container, container.parentNode);
 
         const hW = elWidth / 2,
             hH = elHeight / 2;
@@ -571,15 +558,6 @@ export default class Draggable extends Transformable {
 
         const originTransform = cHandle ? getTransform(cHandle) : createIdentityMatrix();
 
-        const {
-            translate: {
-                x: originX,
-                y: originY
-            }
-        } = cHandle
-            ? decompose(getCurrentTransformMatrix(cHandle))
-            : { translate: { x: globalCenterX, y: globalCenterY } };
-
         // search distance between el's center and rotation handle
         const [distX, distY] = multiplyMatrixAndPoint(
             multiplyMatrix(
@@ -587,8 +565,8 @@ export default class Draggable extends Transformable {
                 dropTranslate(originTransform)
             ),
             [
-                originX - globalCenterX,
-                originY - globalCenterY,
+                transformOrigin[0] - globalCenterX,
+                transformOrigin[1] - globalCenterY,
                 0,
                 1
             ]
@@ -603,11 +581,6 @@ export default class Draggable extends Transformable {
                 0,
                 1
             ]
-        );
-
-        const containerBox = multiplyMatrixAndPoint(
-            dropTranslate(containerMatrix),
-            [offsetWidth, offsetHeight, 0, 1]
         );
 
         const {
@@ -639,11 +612,8 @@ export default class Draggable extends Transformable {
             matrix,
             ctm,
             parentMatrix,
-            containerMatrix,
-            wrapperMatrix,
             scX: sX,
-            scY: sY,
-            containerBox
+            scY: sY
         };
 
         return {
@@ -669,7 +639,8 @@ export default class Draggable extends Transformable {
                 handles: {
                     center: cHandle
                 },
-                center: oldCenter
+                center: oldCenter,
+                wrapper
             },
             options: {
                 container,
@@ -679,6 +650,9 @@ export default class Draggable extends Transformable {
 
         const [glLeft, glTop] = getAbsoluteOffset(elements[0], container);
         const ctm = getCurrentTransformMatrix(elements[0], container);
+
+        const restrictContainer = restrict || container;
+        const containerMatrix = getCurrentTransformMatrix(restrictContainer, restrictContainer.parentNode);
 
         const {
             width: boxWidth,
@@ -704,9 +678,8 @@ export default class Draggable extends Transformable {
         return {
             transform: {
                 controlsMatrix: getCurrentTransformMatrix(controls, controls.parentNode),
-                containerMatrix: restrict
-                    ? getCurrentTransformMatrix(restrict, restrict.parentNode)
-                    : getCurrentTransformMatrix(container, container.parentNode)
+                containerMatrix,
+                wrapperMatrix: getCurrentTransformMatrix(wrapper, container)
             },
             bBox: {
                 ...this._getBBox()
@@ -755,12 +728,8 @@ export default class Draggable extends Transformable {
     }
 
     _processControlsResize() {
-        this._applyTransformToHandles({
-            boxMatrix: createIdentityMatrix()
-        });
-    }
+        const { center } = this._applyTransformToHandles();
 
-    _processControlsMove({ dx, dy }) {
         const {
             storage: {
                 transform: {
@@ -769,40 +738,101 @@ export default class Draggable extends Transformable {
             }
         } = this;
 
+        if (!center) return;
+
+        this.storage = {
+            ...this.storage,
+            cached: {
+                transformOrigin: multiplyMatrixAndPoint(
+                    controlsMatrix,
+                    center
+                )
+            }
+        };
+    }
+
+    _processControlsMove({ dx, dy }) {
+        const {
+            storage: {
+                transform: {
+                    controlsMatrix,
+                    wrapperMatrix
+                },
+                center,
+                transformOrigin
+            }
+        } = this;
+
         const moveControlsMtrx = multiplyMatrix(
             controlsMatrix,
             createTranslateMatrix(dx, dy)
         );
 
-        const wrapperStyle = matrixToCSS(flatMatrix(moveControlsMtrx));
+        this._updateControlsView(moveControlsMtrx);
 
-        this._updateControlsView(wrapperStyle);
+        const centerTransformMatrix = dropTranslate(matrixInvert(wrapperMatrix));
+        const [cx, cy] = multiplyMatrixAndPoint(
+            centerTransformMatrix,
+            [dx, dy, 0, 1]
+        );
+
+        if (center.isShifted) {
+            this._moveCenterHandle(-cx, -cy, false);
+        } else {
+            const translateMatrix = createTranslateMatrix(cx, cy);
+
+            this.storage = {
+                ...this.storage,
+                cached: {
+                    transformOrigin: multiplyMatrixAndPoint(
+                        translateMatrix,
+                        transformOrigin
+                    )
+                }
+            };
+        }
     }
 
     _processControlsRotate() {
-        this._applyTransformToHandles({
-            boxMatrix: createIdentityMatrix()
-        });
+        this._applyTransformToHandles();
     }
 
-    _moveCenterHandle(x, y) {
+    _moveCenterHandle(x, y, updateTransformOrigin = true) {
         const {
             storage: {
                 handles: { center },
-                center: { matrix }
+                center: { matrix },
+                center: prevCenterData,
+                transformOrigin
             }
         } = this;
 
+        const translateMatrix = createTranslateMatrix(x, y);
+
         const resultMatrix = multiplyMatrix(
             matrix,
-            createTranslateMatrix(x, y)
+            translateMatrix
         );
 
         helper(center).css({
             ...matrixToCSS(flatMatrix(resultMatrix))
         });
 
-        this.storage.center.isShifted = true;
+        this.storage = {
+            ...this.storage,
+            center: {
+                ...prevCenterData,
+                isShifted: true
+            },
+            ...(updateTransformOrigin ? {
+                cached: {
+                    transformOrigin: multiplyMatrixAndPoint(
+                        translateMatrix,
+                        transformOrigin
+                    )
+                }
+            } : {})
+        };
     }
 
     _processMoveRestrict(element, { dx, dy }) {
@@ -961,8 +991,11 @@ export default class Draggable extends Transformable {
         helper(element).css(css);
     }
 
-    _updateControlsView(css) {
-        helper(this.storage.controls).css(css);
+    _updateControlsView(matrix = createIdentityMatrix()) {
+        const cssStyle = matrixToCSS(flatMatrix(matrix));
+        helper(this.storage.controls).css(cssStyle);
+
+        this.storage.cached.controlsMatrix = matrix;
     }
 
     _getVertices(transformMatrix = createIdentityMatrix()) {
@@ -1227,10 +1260,11 @@ export default class Draggable extends Transformable {
             },
             storage: {
                 handles,
+                controls,
                 center: { isShifted = false } = {},
                 transform: {
-                    controlsMatrix
-                }
+                    controlsMatrix = getCurrentTransformMatrix(controls, controls.parentNode)
+                } = {}
             }
         } = this;
 
@@ -1288,65 +1322,120 @@ export default class Draggable extends Transformable {
             ...((!isShifted && Boolean(center)) && { center })
         };
 
-        keys(allHandles).forEach(key => {
+        return keys(allHandles).reduce((result, key) => {
             const hdl = handles[key];
             const attr = allHandles[key];
-            if (isUndef(attr) || isUndef(hdl)) return;
+
+            result[key] = attr;
+
+            if (isUndef(attr) || isUndef(hdl)) return result;
 
             const [x, y] = attr;
             helper(hdl).css({
                 transform: `translate(${x}px, ${y}px)`
             });
-        });
+
+            return result;
+        }, {});
     }
 
-    resetCenterPoint() {
+    setCenterPoint(...args) {
+        warn('"setCenterPoint" method is replaced by "setTransformOrigin" and would be removed soon');
+        this.setTransformOrigin(...args);
+    }
+
+    setTransformOrigin({ x, y, dx, dy } = {}, pin = true) {
         const {
             elements: [element] = [],
+            storage,
             storage: {
                 wrapper,
-                handles: { center }
-            },
+                handles: {
+                    center: handle
+                },
+                center
+            } = {},
             options: {
                 container
-            }
+            } = {}
         } = this;
 
-        if (!center) return;
+        const isRelative = isDef(dx) && isDef(dy),
+            isAbsolute = isDef(x) && isDef(y);
 
-        const { offsetHeight, offsetWidth } = element;
-
-        const [offsetLeft, offsetTop] = getAbsoluteOffset(element, container);
+        if (!handle || !center || !(isRelative || isAbsolute)) return;
 
         const matrix = multiplyMatrix(
             getCurrentTransformMatrix(element, container),
             matrixInvert(getCurrentTransformMatrix(wrapper, wrapper.parentNode))
         );
 
-        const [x, y] = multiplyMatrixAndPoint(
-            matrix,
-            [offsetWidth / 2, offsetHeight / 2, 0, 1]
-        );
+        let newX, newY;
 
-        helper(center).css(
-            { transform: `translate(${x + offsetLeft}px, ${y + offsetTop}px)` }
+        const [offsetLeft, offsetTop] = getAbsoluteOffset(element, container);
+
+        if (isRelative) {
+            const { offsetHeight, offsetWidth } = element;
+
+            const relX = -dx + offsetWidth / 2;
+            const relY = -dy + offsetHeight / 2;
+
+            [newX, newY] = multiplyMatrixAndPoint(
+                matrix,
+                [relX, relY, 0, 1]
+            );
+        } else {
+            newX = x;
+            newY = y;
+        }
+
+        helper(handle).css({
+            transform: `translate(${newX + offsetLeft}px, ${newY + offsetTop}px)`
+        });
+
+        center.isShifted = pin;
+        storage.transformOrigin = multiplyMatrixAndPoint(
+            createIdentityMatrix(),
+            [newX, newY, 0, 1]
         );
     }
 
     fitControlsToSize() {
-        const identityMatrix = createIdentityMatrix();
-
-        this.storage = {
-            ...this.storage,
-            transform: {
-                ...(this.storage.transform || {}),
-                controlsMatrix: identityMatrix
+        const {
+            storage: {
+                controls,
+                center: {
+                    isShifted
+                } = {},
+                transformOrigin: [
+                    originX,
+                    originY
+                ]
             }
-        };
+        } = this;
 
-        this._updateControlsView(
-            matrixToCSS(flatMatrix(identityMatrix))
+        const controlsMatrix = getCurrentTransformMatrix(controls, controls.parentNode);
+        const [dx, dy] =  multiplyMatrixAndPoint(
+            controlsMatrix,
+            [originX, originY, 0, 1]
         );
+
+        const { nextValues, pin } = [
+            {
+                nextValues: () => ({ x: dx, y: dy }),
+                pin: true,
+                condition: () => isShifted
+            },
+            {
+                nextValues: () => ({ dx: 0, dy: 0 }),
+                pin: false,
+                condition: () => !isShifted
+            }
+        ].find(({ condition }) => condition());
+
+        this._updateControlsView();
+
+        this.setTransformOrigin({ ...nextValues() }, pin);
         this._applyTransformToHandles();
     }
 
@@ -1459,11 +1548,35 @@ export default class Draggable extends Transformable {
             createTranslateMatrix(x, y)
         );
 
-        this._updateElementView(
-            elements[0],
-            matrixToCSS(flatMatrix(moveElementMtrx))
-        );
+        elements.map((element) => (
+            this._updateElementView(
+                element,
+                matrixToCSS(flatMatrix(moveElementMtrx))
+            )
+        ));
+
         this.fitControlsToSize();
+    }
+
+    getDimensions() {
+        const {
+            elements: [element] = [],
+            options: {
+                isGrouped
+            }
+        } = this;
+
+        const { tl, tr, br } = isGrouped
+            ? this._getGroupVertices()
+            : this._getElementVertices(element, createIdentityMatrix());
+
+        return {
+            x: floatToFixed(tl[0]),
+            y: floatToFixed(tl[1]),
+            width: floatToFixed(Math.sqrt(Math.pow(tl[0] - tr[0], 2) + Math.pow(tl[1] - tr[1], 2))),
+            height: floatToFixed(Math.sqrt(Math.pow(tr[0] - br[0], 2) + Math.pow(tr[1] - br[1], 2))),
+            rotation: floatToFixed(Math.atan2(tr[1] - tl[1], tr[0] - tl[0]) * DEG)
+        };
     }
 
 }