+++ CHANGELOG.md
@@ -1,3 +1,54 @@
+# [1.1.0-rc.4](https://github.com/nichollascarter/subjx/compare/v1.1.0-rc.3...v1.1.0-rc.4) (2022-03-21)
+
+
+### Bug Fixes
+
+* **core:** fixed tests ([5d81abe](https://github.com/nichollascarter/subjx/commit/5d81abe94f745b25d4d539d081bcd14a9870003a))
+* **core:** set correct transform origin, allow alignment for groupable html elements ([cf29d29](https://github.com/nichollascarter/subjx/commit/cf29d2910cbea89c83864b98d382cbfdda535531))
+
+# [1.1.0-rc.3](https://github.com/nichollascarter/subjx/compare/v1.1.0-rc.2...v1.1.0-rc.3) (2022-03-02)
+
+
+### Bug Fixes
+
+* **src:** fixed undefined containerMatrix value on applying alignment ([6b20349](https://github.com/nichollascarter/subjx/commit/6b20349b3ca408297d923f64646bc69ac860245c))
+
+# [1.1.0-rc.2](https://github.com/nichollascarter/subjx/compare/v1.1.0-rc.1...v1.1.0-rc.2) (2022-02-14)
+
+
+### Bug Fixes
+
+* fix wrong transform-origin position for nested elements ([93af678](https://github.com/nichollascarter/subjx/commit/93af67887834b0c455ab8cbb38b8d7a7765d9ae5))
+
+# [1.1.0-rc.2](https://github.com/nichollascarter/subjx/compare/v1.1.0-rc.1...v1.1.0-rc.2) (2022-02-14)
+
+
+### Bug Fixes
+
+* fix wrong transform-origin position for nested elements ([93af678](https://github.com/nichollascarter/subjx/commit/93af67887834b0c455ab8cbb38b8d7a7765d9ae5))
+
+# [1.1.0-rc.2](https://github.com/nichollascarter/subjx/compare/v1.1.0-rc.1...v1.1.0-rc.2) (2022-02-14)
+
+
+### Bug Fixes
+
+* fix wrong transform-origin position for nested elements ([93af678](https://github.com/nichollascarter/subjx/commit/93af67887834b0c455ab8cbb38b8d7a7765d9ae5))
+
+# [1.1.0-rc.2](https://github.com/nichollascarter/subjx/compare/v1.1.0-rc.1...v1.1.0-rc.2) (2022-02-14)
+
+
+### Bug Fixes
+
+* fix wrong transform-origin position for nested elements ([93af678](https://github.com/nichollascarter/subjx/commit/93af67887834b0c455ab8cbb38b8d7a7765d9ae5))
+
+# [1.1.0-rc.1](https://github.com/nichollascarter/subjx/compare/v1.0.0...v1.1.0-rc.1) (2022-02-13)
+
+
+### Features
+
+* **src:** add setCenterPoint method ([7995cf5](https://github.com/nichollascarter/subjx/commit/7995cf504434d98e60bfa6e240a7e14eee8372fb)), closes [#56](https://github.com/nichollascarter/subjx/issues/56)
+* added transform origin support ([75e33de](https://github.com/nichollascarter/subjx/commit/75e33de273d896b3b3e64593be123bea4dd6d64f))
+
 # [1.0.0](https://github.com/nichollascarter/subjx/compare/v0.3.9...v1.0.0) (2021-12-15)